@@ -2,18 +2,21 @@ package data
 
 import (
 	"context"
+	v1 "review-b/api/review/v1" // 添加这行导入
 	"review-b/internal/conf"
-	v1 "review-b/api/review/v1"  // 添加这行导入
 
+	"github.com/go-kratos/kratos/contrib/registry/consul/v2"
 	"github.com/go-kratos/kratos/v2/log"
 	"github.com/go-kratos/kratos/v2/middleware/recovery"
 	"github.com/go-kratos/kratos/v2/middleware/validate"
+	"github.com/go-kratos/kratos/v2/registry"
 	"github.com/go-kratos/kratos/v2/transport/grpc"
 	"github.com/google/wire"
+	"github.com/hashicorp/consul/api"
 )
 
 // ProviderSet is data providers.
-var ProviderSet = wire.NewSet(NewReviewServiceClient, NewData, NewBusinessRepo)
+var ProviderSet = wire.NewSet(NewDiscovery, NewReviewServiceClient, NewData, NewBusinessRepo)
 
 // Data .
 type Data struct {
@@ -34,12 +37,30 @@ func NewData(c *conf.Data, rc v1.ReviewClient, logger log.Logger) (*Data, func()
 	}, cleanup, nil
 }
 
+// NewDiscovery 服务发现对象的构造函数
+func NewDiscovery(conf *conf.Registry) registry.Discovery {
+	// new consul client
+	c := api.DefaultConfig()
+	c.Address = conf.Consul.Address
+	c.Scheme = conf.Consul.Scheme
+
+	client, err := api.NewClient(c)
+	if err != nil {
+		panic(err)
+	}
+	// new reg with consul client
+	dis := consul.New(client)
+	return dis
+}
+
 // NewReviewServiceClient 创建一个连接 review-service 的gRPC Client 端
-func NewReviewServiceClient() v1.ReviewClient {
+func NewReviewServiceClient(d registry.Discovery) v1.ReviewClient {
 	// import "github.com/go-kratos/kratos/v2/transport/grpc"
 	conn, err := grpc.DialInsecure(
 		context.Background(),
-		grpc.WithEndpoint("127.0.0.1:9092"),
+		// grpc.WithEndpoint("127.0.0.1:9092"),
+		grpc.WithEndpoint("discovery:///review.service"),
+		grpc.WithDiscovery(d), // 服务发现
 		grpc.WithMiddleware(
 			recovery.Recovery(),
 			validate.Validator(),
