@@ -1,8 +1,8 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.8
-// 	protoc        v6.30.2
-// source: api/review/v1/review_error.proto
+// 	protoc-gen-go v1.31.0
+// 	protoc        v3.20.1
+// source: review/v1/review_error.proto
 
 package v1
 
@@ -12,7 +12,6 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
-	unsafe "unsafe"
 )
 
 const (
@@ -56,11 +55,11 @@ func (x ErrorReason) String() string {
 }
 
 func (ErrorReason) Descriptor() protoreflect.EnumDescriptor {
-	return file_api_review_v1_review_error_proto_enumTypes[0].Descriptor()
+	return file_review_v1_review_error_proto_enumTypes[0].Descriptor()
 }
 
 func (ErrorReason) Type() protoreflect.EnumType {
-	return &file_api_review_v1_review_error_proto_enumTypes[0]
+	return &file_review_v1_review_error_proto_enumTypes[0]
 }
 
 func (x ErrorReason) Number() protoreflect.EnumNumber {
@@ -69,38 +68,45 @@ func (x ErrorReason) Number() protoreflect.EnumNumber {
 
 // Deprecated: Use ErrorReason.Descriptor instead.
 func (ErrorReason) EnumDescriptor() ([]byte, []int) {
-	return file_api_review_v1_review_error_proto_rawDescGZIP(), []int{0}
+	return file_review_v1_review_error_proto_rawDescGZIP(), []int{0}
 }
 
-var File_api_review_v1_review_error_proto protoreflect.FileDescriptor
+var File_review_v1_review_error_proto protoreflect.FileDescriptor
 
-const file_api_review_v1_review_error_proto_rawDesc = "" +
-	"\n" +
-	" api/review/v1/review_error.proto\x12\rapi.review.v1\x1a\x13errors/errors.proto*X\n" +
-	"\vErrorReason\x12\x14\n" +
-	"\n" +
-	"NEED_LOGIN\x10\x00\x1a\x04\xa8E\x91\x03\x12\x13\n" +
-	"\tDB_FAILED\x10\x01\x1a\x04\xa8E\xf4\x03\x12\x18\n" +
-	"\x0eORDER_REVIEWED\x10d\x1a\x04\xa8E\x90\x03\x1a\x04\xa0E\xf4\x03B2\n" +
-	"\rapi.review.v1P\x01Z\x1freview-service/api/review/v1;v1b\x06proto3"
+var file_review_v1_review_error_proto_rawDesc = []byte{
+	0x0a, 0x1c, 0x72, 0x65, 0x76, 0x69, 0x65, 0x77, 0x2f, 0x76, 0x31, 0x2f, 0x72, 0x65, 0x76, 0x69,
+	0x65, 0x77, 0x5f, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x0d,
+	0x61, 0x70, 0x69, 0x2e, 0x72, 0x65, 0x76, 0x69, 0x65, 0x77, 0x2e, 0x76, 0x31, 0x1a, 0x13, 0x65,
+	0x72, 0x72, 0x6f, 0x72, 0x73, 0x2f, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x73, 0x2e, 0x70, 0x72, 0x6f,
+	0x74, 0x6f, 0x2a, 0x58, 0x0a, 0x0b, 0x45, 0x72, 0x72, 0x6f, 0x72, 0x52, 0x65, 0x61, 0x73, 0x6f,
+	0x6e, 0x12, 0x14, 0x0a, 0x0a, 0x4e, 0x45, 0x45, 0x44, 0x5f, 0x4c, 0x4f, 0x47, 0x49, 0x4e, 0x10,
+	0x00, 0x1a, 0x04, 0xa8, 0x45, 0x91, 0x03, 0x12, 0x13, 0x0a, 0x09, 0x44, 0x42, 0x5f, 0x46, 0x41,
+	0x49, 0x4c, 0x45, 0x44, 0x10, 0x01, 0x1a, 0x04, 0xa8, 0x45, 0xf4, 0x03, 0x12, 0x18, 0x0a, 0x0e,
+	0x4f, 0x52, 0x44, 0x45, 0x52, 0x5f, 0x52, 0x45, 0x56, 0x49, 0x45, 0x57, 0x45, 0x44, 0x10, 0x64,
+	0x1a, 0x04, 0xa8, 0x45, 0x90, 0x03, 0x1a, 0x04, 0xa0, 0x45, 0xf4, 0x03, 0x42, 0x32, 0x0a, 0x0d,
+	0x61, 0x70, 0x69, 0x2e, 0x72, 0x65, 0x76, 0x69, 0x65, 0x77, 0x2e, 0x76, 0x31, 0x50, 0x01, 0x5a,
+	0x1f, 0x72, 0x65, 0x76, 0x69, 0x65, 0x77, 0x2d, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x2f,
+	0x61, 0x70, 0x69, 0x2f, 0x72, 0x65, 0x76, 0x69, 0x65, 0x77, 0x2f, 0x76, 0x31, 0x3b, 0x76, 0x31,
+	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
+}
 
 var (
-	file_api_review_v1_review_error_proto_rawDescOnce sync.Once
-	file_api_review_v1_review_error_proto_rawDescData []byte
+	file_review_v1_review_error_proto_rawDescOnce sync.Once
+	file_review_v1_review_error_proto_rawDescData = file_review_v1_review_error_proto_rawDesc
 )
 
-func file_api_review_v1_review_error_proto_rawDescGZIP() []byte {
-	file_api_review_v1_review_error_proto_rawDescOnce.Do(func() {
-		file_api_review_v1_review_error_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_api_review_v1_review_error_proto_rawDesc), len(file_api_review_v1_review_error_proto_rawDesc)))
+func file_review_v1_review_error_proto_rawDescGZIP() []byte {
+	file_review_v1_review_error_proto_rawDescOnce.Do(func() {
+		file_review_v1_review_error_proto_rawDescData = protoimpl.X.CompressGZIP(file_review_v1_review_error_proto_rawDescData)
 	})
-	return file_api_review_v1_review_error_proto_rawDescData
+	return file_review_v1_review_error_proto_rawDescData
 }
 
-var file_api_review_v1_review_error_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
-var file_api_review_v1_review_error_proto_goTypes = []any{
+var file_review_v1_review_error_proto_enumTypes = make([]protoimpl.EnumInfo, 1)
+var file_review_v1_review_error_proto_goTypes = []interface{}{
 	(ErrorReason)(0), // 0: api.review.v1.ErrorReason
 }
-var file_api_review_v1_review_error_proto_depIdxs = []int32{
+var file_review_v1_review_error_proto_depIdxs = []int32{
 	0, // [0:0] is the sub-list for method output_type
 	0, // [0:0] is the sub-list for method input_type
 	0, // [0:0] is the sub-list for extension type_name
@@ -108,26 +114,27 @@ var file_api_review_v1_review_error_proto_depIdxs = []int32{
 	0, // [0:0] is the sub-list for field type_name
 }
 
-func init() { file_api_review_v1_review_error_proto_init() }
-func file_api_review_v1_review_error_proto_init() {
-	if File_api_review_v1_review_error_proto != nil {
+func init() { file_review_v1_review_error_proto_init() }
+func file_review_v1_review_error_proto_init() {
+	if File_review_v1_review_error_proto != nil {
 		return
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: unsafe.Slice(unsafe.StringData(file_api_review_v1_review_error_proto_rawDesc), len(file_api_review_v1_review_error_proto_rawDesc)),
+			RawDescriptor: file_review_v1_review_error_proto_rawDesc,
 			NumEnums:      1,
 			NumMessages:   0,
 			NumExtensions: 0,
 			NumServices:   0,
 		},
-		GoTypes:           file_api_review_v1_review_error_proto_goTypes,
-		DependencyIndexes: file_api_review_v1_review_error_proto_depIdxs,
-		EnumInfos:         file_api_review_v1_review_error_proto_enumTypes,
+		GoTypes:           file_review_v1_review_error_proto_goTypes,
+		DependencyIndexes: file_review_v1_review_error_proto_depIdxs,
+		EnumInfos:         file_review_v1_review_error_proto_enumTypes,
 	}.Build()
-	File_api_review_v1_review_error_proto = out.File
-	file_api_review_v1_review_error_proto_goTypes = nil
-	file_api_review_v1_review_error_proto_depIdxs = nil
+	File_review_v1_review_error_proto = out.File
+	file_review_v1_review_error_proto_rawDesc = nil
+	file_review_v1_review_error_proto_goTypes = nil
+	file_review_v1_review_error_proto_depIdxs = nil
 }
