@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.5.1
-// - protoc             v6.30.2
+// - protoc-gen-go-grpc v1.3.0
+// - protoc             v3.20.1
 // source: review/v1/review.proto
 
 package v1
@@ -15,24 +15,23 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.64.0 or later.
-const _ = grpc.SupportPackageIsVersion9
+// Requires gRPC-Go v1.32.0 or later.
+const _ = grpc.SupportPackageIsVersion7
 
 const (
-	Review_CreateReview_FullMethodName       = "/api.review.v1.Review/CreateReview"
-	Review_GetReview_FullMethodName          = "/api.review.v1.Review/GetReview"
-	Review_AuditReview_FullMethodName        = "/api.review.v1.Review/AuditReview"
-	Review_ReplyReview_FullMethodName        = "/api.review.v1.Review/ReplyReview"
-	Review_AppealReview_FullMethodName       = "/api.review.v1.Review/AppealReview"
-	Review_AuditAppeal_FullMethodName        = "/api.review.v1.Review/AuditAppeal"
-	Review_ListReviewByUserID_FullMethodName = "/api.review.v1.Review/ListReviewByUserID"
+	Review_CreateReview_FullMethodName        = "/api.review.v1.Review/CreateReview"
+	Review_GetReview_FullMethodName           = "/api.review.v1.Review/GetReview"
+	Review_AuditReview_FullMethodName         = "/api.review.v1.Review/AuditReview"
+	Review_ReplyReview_FullMethodName         = "/api.review.v1.Review/ReplyReview"
+	Review_AppealReview_FullMethodName        = "/api.review.v1.Review/AppealReview"
+	Review_AuditAppeal_FullMethodName         = "/api.review.v1.Review/AuditAppeal"
+	Review_ListReviewByUserID_FullMethodName  = "/api.review.v1.Review/ListReviewByUserID"
+	Review_ListReviewByStoreID_FullMethodName = "/api.review.v1.Review/ListReviewByStoreID"
 )
 
 // ReviewClient is the client API for Review service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
-//
-// 定义评价服务
 type ReviewClient interface {
 	// C端创建评价
 	CreateReview(ctx context.Context, in *CreateReviewRequest, opts ...grpc.CallOption) (*CreateReviewReply, error)
@@ -48,6 +47,8 @@ type ReviewClient interface {
 	AuditAppeal(ctx context.Context, in *AuditAppealRequest, opts ...grpc.CallOption) (*AuditAppealReply, error)
 	// C端查看userID下所有评价
 	ListReviewByUserID(ctx context.Context, in *ListReviewByUserIDRequest, opts ...grpc.CallOption) (*ListReviewByUserIDReply, error)
+	// 根据商家ID查询评价列表（分页）
+	ListReviewByStoreID(ctx context.Context, in *ListReviewByStoreIDRequest, opts ...grpc.CallOption) (*ListReviewByStoreIDReply, error)
 }
 
 type reviewClient struct {
@@ -59,9 +60,8 @@ func NewReviewClient(cc grpc.ClientConnInterface) ReviewClient {
 }
 
 func (c *reviewClient) CreateReview(ctx context.Context, in *CreateReviewRequest, opts ...grpc.CallOption) (*CreateReviewReply, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreateReviewReply)
-	err := c.cc.Invoke(ctx, Review_CreateReview_FullMethodName, in, out, cOpts...)
+	err := c.cc.Invoke(ctx, Review_CreateReview_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -69,9 +69,8 @@ func (c *reviewClient) CreateReview(ctx context.Context, in *CreateReviewRequest
 }
 
 func (c *reviewClient) GetReview(ctx context.Context, in *GetReviewRequest, opts ...grpc.CallOption) (*GetReviewReply, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GetReviewReply)
-	err := c.cc.Invoke(ctx, Review_GetReview_FullMethodName, in, out, cOpts...)
+	err := c.cc.Invoke(ctx, Review_GetReview_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -79,9 +78,8 @@ func (c *reviewClient) GetReview(ctx context.Context, in *GetReviewRequest, opts
 }
 
 func (c *reviewClient) AuditReview(ctx context.Context, in *AuditReviewRequest, opts ...grpc.CallOption) (*AuditReviewReply, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AuditReviewReply)
-	err := c.cc.Invoke(ctx, Review_AuditReview_FullMethodName, in, out, cOpts...)
+	err := c.cc.Invoke(ctx, Review_AuditReview_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -89,9 +87,8 @@ func (c *reviewClient) AuditReview(ctx context.Context, in *AuditReviewRequest, 
 }
 
 func (c *reviewClient) ReplyReview(ctx context.Context, in *ReplyReviewRequest, opts ...grpc.CallOption) (*ReplyReviewReply, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ReplyReviewReply)
-	err := c.cc.Invoke(ctx, Review_ReplyReview_FullMethodName, in, out, cOpts...)
+	err := c.cc.Invoke(ctx, Review_ReplyReview_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -99,9 +96,8 @@ func (c *reviewClient) ReplyReview(ctx context.Context, in *ReplyReviewRequest, 
 }
 
 func (c *reviewClient) AppealReview(ctx context.Context, in *AppealReviewRequest, opts ...grpc.CallOption) (*AppealReviewReply, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AppealReviewReply)
-	err := c.cc.Invoke(ctx, Review_AppealReview_FullMethodName, in, out, cOpts...)
+	err := c.cc.Invoke(ctx, Review_AppealReview_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -109,9 +105,8 @@ func (c *reviewClient) AppealReview(ctx context.Context, in *AppealReviewRequest
 }
 
 func (c *reviewClient) AuditAppeal(ctx context.Context, in *AuditAppealRequest, opts ...grpc.CallOption) (*AuditAppealReply, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(AuditAppealReply)
-	err := c.cc.Invoke(ctx, Review_AuditAppeal_FullMethodName, in, out, cOpts...)
+	err := c.cc.Invoke(ctx, Review_AuditAppeal_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -119,9 +114,17 @@ func (c *reviewClient) AuditAppeal(ctx context.Context, in *AuditAppealRequest, 
 }
 
 func (c *reviewClient) ListReviewByUserID(ctx context.Context, in *ListReviewByUserIDRequest, opts ...grpc.CallOption) (*ListReviewByUserIDReply, error) {
-	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListReviewByUserIDReply)
-	err := c.cc.Invoke(ctx, Review_ListReviewByUserID_FullMethodName, in, out, cOpts...)
+	err := c.cc.Invoke(ctx, Review_ListReviewByUserID_FullMethodName, in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *reviewClient) ListReviewByStoreID(ctx context.Context, in *ListReviewByStoreIDRequest, opts ...grpc.CallOption) (*ListReviewByStoreIDReply, error) {
+	out := new(ListReviewByStoreIDReply)
+	err := c.cc.Invoke(ctx, Review_ListReviewByStoreID_FullMethodName, in, out, opts...)
 	if err != nil {
 		return nil, err
 	}
@@ -130,9 +133,7 @@ func (c *reviewClient) ListReviewByUserID(ctx context.Context, in *ListReviewByU
 
 // ReviewServer is the server API for Review service.
 // All implementations must embed UnimplementedReviewServer
-// for forward compatibility.
-//
-// 定义评价服务
+// for forward compatibility
 type ReviewServer interface {
 	// C端创建评价
 	CreateReview(context.Context, *CreateReviewRequest) (*CreateReviewReply, error)
@@ -148,15 +149,14 @@ type ReviewServer interface {
 	AuditAppeal(context.Context, *AuditAppealRequest) (*AuditAppealReply, error)
 	// C端查看userID下所有评价
 	ListReviewByUserID(context.Context, *ListReviewByUserIDRequest) (*ListReviewByUserIDReply, error)
+	// 根据商家ID查询评价列表（分页）
+	ListReviewByStoreID(context.Context, *ListReviewByStoreIDRequest) (*ListReviewByStoreIDReply, error)
 	mustEmbedUnimplementedReviewServer()
 }
 
-// UnimplementedReviewServer must be embedded to have
-// forward compatible implementations.
-//
-// NOTE: this should be embedded by value instead of pointer to avoid a nil
-// pointer dereference when methods are called.
-type UnimplementedReviewServer struct{}
+// UnimplementedReviewServer must be embedded to have forward compatible implementations.
+type UnimplementedReviewServer struct {
+}
 
 func (UnimplementedReviewServer) CreateReview(context.Context, *CreateReviewRequest) (*CreateReviewReply, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method CreateReview not implemented")
@@ -179,8 +179,10 @@ func (UnimplementedReviewServer) AuditAppeal(context.Context, *AuditAppealReques
 func (UnimplementedReviewServer) ListReviewByUserID(context.Context, *ListReviewByUserIDRequest) (*ListReviewByUserIDReply, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListReviewByUserID not implemented")
 }
+func (UnimplementedReviewServer) ListReviewByStoreID(context.Context, *ListReviewByStoreIDRequest) (*ListReviewByStoreIDReply, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method ListReviewByStoreID not implemented")
+}
 func (UnimplementedReviewServer) mustEmbedUnimplementedReviewServer() {}
-func (UnimplementedReviewServer) testEmbeddedByValue()                {}
 
 // UnsafeReviewServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ReviewServer will
@@ -190,13 +192,6 @@ type UnsafeReviewServer interface {
 }
 
 func RegisterReviewServer(s grpc.ServiceRegistrar, srv ReviewServer) {
-	// If the following call pancis, it indicates UnimplementedReviewServer was
-	// embedded by pointer and is nil.  This will cause panics if an
-	// unimplemented method is ever invoked, so we test this at initialization
-	// time to prevent it from happening at runtime later due to I/O.
-	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
-		t.testEmbeddedByValue()
-	}
 	s.RegisterService(&Review_ServiceDesc, srv)
 }
 
@@ -326,6 +321,24 @@ func _Review_ListReviewByUserID_Handler(srv interface{}, ctx context.Context, de
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Review_ListReviewByStoreID_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(ListReviewByStoreIDRequest)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ReviewServer).ListReviewByStoreID(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: Review_ListReviewByStoreID_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ReviewServer).ListReviewByStoreID(ctx, req.(*ListReviewByStoreIDRequest))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 // Review_ServiceDesc is the grpc.ServiceDesc for Review service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -360,6 +373,10 @@ var Review_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "ListReviewByUserID",
 			Handler:    _Review_ListReviewByUserID_Handler,
+		},
+		{
+			MethodName: "ListReviewByStoreID",
+			Handler:    _Review_ListReviewByStoreID_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
