@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-http. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-http v2.8.4
-// - protoc             v6.30.2
+// - protoc-gen-go-http v2.7.0
+// - protoc             v3.20.1
 // source: review/v1/review.proto
 
 package v1
@@ -48,10 +48,10 @@ func RegisterReviewHTTPServer(s *http.Server, srv ReviewHTTPServer) {
 	r := s.Route("/")
 	r.POST("/v1/review", _Review_CreateReview0_HTTP_Handler(srv))
 	r.GET("/v1/review/{reviewID}", _Review_GetReview0_HTTP_Handler(srv))
-	r.POST("/v1/review/audit", _Review_AuditReview0_HTTP_Handler(srv))
-	r.POST("/v1/review/reply", _Review_ReplyReview0_HTTP_Handler(srv))
-	r.POST("/v1/review/appeal", _Review_AppealReview0_HTTP_Handler(srv))
-	r.POST("/v1/appeal/audit", _Review_AuditAppeal0_HTTP_Handler(srv))
+	r.POST("/v1/review/audit", _Review_AuditReview1_HTTP_Handler(srv))
+	r.POST("/v1/review/reply", _Review_ReplyReview1_HTTP_Handler(srv))
+	r.POST("/v1/review/appeal", _Review_AppealReview1_HTTP_Handler(srv))
+	r.POST("/v1/appeal/audit", _Review_AuditAppeal1_HTTP_Handler(srv))
 	r.GET("/v1/{userID}/reviews", _Review_ListReviewByUserID0_HTTP_Handler(srv))
 }
 
@@ -99,7 +99,7 @@ func _Review_GetReview0_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context
 	}
 }
 
-func _Review_AuditReview0_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context) error {
+func _Review_AuditReview1_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context) error {
 	return func(ctx http.Context) error {
 		var in AuditReviewRequest
 		if err := ctx.Bind(&in); err != nil {
@@ -121,7 +121,7 @@ func _Review_AuditReview0_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Conte
 	}
 }
 
-func _Review_ReplyReview0_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context) error {
+func _Review_ReplyReview1_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context) error {
 	return func(ctx http.Context) error {
 		var in ReplyReviewRequest
 		if err := ctx.Bind(&in); err != nil {
@@ -143,7 +143,7 @@ func _Review_ReplyReview0_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Conte
 	}
 }
 
-func _Review_AppealReview0_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context) error {
+func _Review_AppealReview1_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context) error {
 	return func(ctx http.Context) error {
 		var in AppealReviewRequest
 		if err := ctx.Bind(&in); err != nil {
@@ -165,7 +165,7 @@ func _Review_AppealReview0_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Cont
 	}
 }
 
-func _Review_AuditAppeal0_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context) error {
+func _Review_AuditAppeal1_HTTP_Handler(srv ReviewHTTPServer) func(ctx http.Context) error {
 	return func(ctx http.Context) error {
 		var in AuditAppealRequest
 		if err := ctx.Bind(&in); err != nil {
@@ -237,7 +237,7 @@ func (c *ReviewHTTPClientImpl) AppealReview(ctx context.Context, in *AppealRevie
 	if err != nil {
 		return nil, err
 	}
-	return &out, nil
+	return &out, err
 }
 
 func (c *ReviewHTTPClientImpl) AuditAppeal(ctx context.Context, in *AuditAppealRequest, opts ...http.CallOption) (*AuditAppealReply, error) {
@@ -250,7 +250,7 @@ func (c *ReviewHTTPClientImpl) AuditAppeal(ctx context.Context, in *AuditAppealR
 	if err != nil {
 		return nil, err
 	}
-	return &out, nil
+	return &out, err
 }
 
 func (c *ReviewHTTPClientImpl) AuditReview(ctx context.Context, in *AuditReviewRequest, opts ...http.CallOption) (*AuditReviewReply, error) {
@@ -263,7 +263,7 @@ func (c *ReviewHTTPClientImpl) AuditReview(ctx context.Context, in *AuditReviewR
 	if err != nil {
 		return nil, err
 	}
-	return &out, nil
+	return &out, err
 }
 
 func (c *ReviewHTTPClientImpl) CreateReview(ctx context.Context, in *CreateReviewRequest, opts ...http.CallOption) (*CreateReviewReply, error) {
@@ -276,7 +276,7 @@ func (c *ReviewHTTPClientImpl) CreateReview(ctx context.Context, in *CreateRevie
 	if err != nil {
 		return nil, err
 	}
-	return &out, nil
+	return &out, err
 }
 
 func (c *ReviewHTTPClientImpl) GetReview(ctx context.Context, in *GetReviewRequest, opts ...http.CallOption) (*GetReviewReply, error) {
@@ -289,7 +289,7 @@ func (c *ReviewHTTPClientImpl) GetReview(ctx context.Context, in *GetReviewReque
 	if err != nil {
 		return nil, err
 	}
-	return &out, nil
+	return &out, err
 }
 
 func (c *ReviewHTTPClientImpl) ListReviewByUserID(ctx context.Context, in *ListReviewByUserIDRequest, opts ...http.CallOption) (*ListReviewByUserIDReply, error) {
@@ -302,7 +302,7 @@ func (c *ReviewHTTPClientImpl) ListReviewByUserID(ctx context.Context, in *ListR
 	if err != nil {
 		return nil, err
 	}
-	return &out, nil
+	return &out, err
 }
 
 func (c *ReviewHTTPClientImpl) ReplyReview(ctx context.Context, in *ReplyReviewRequest, opts ...http.CallOption) (*ReplyReviewReply, error) {
@@ -315,5 +315,5 @@ func (c *ReviewHTTPClientImpl) ReplyReview(ctx context.Context, in *ReplyReviewR
 	if err != nil {
 		return nil, err
 	}
-	return &out, nil
+	return &out, err
 }
